Split oversized test fixtures
