+++ app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
@@ -2025,103 +2025,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
     }
 
     @Test
-    fun writeInputToPaneIssuesSendKeysWithLiteralBytes() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // Submission semantics: a `\r` byte (carriage return) is the
-        // "Enter / submit" signal coming from the terminal emulator
-        // and from the in-app callers (chips, snippets with-Enter,
-        // composer send) — see [TmuxSessionScreen]. The single-line
-        // route keeps the existing two-token send-keys shape so
-        // keyboard typing still submits cleanly.
-        //
-        // A literal `\n` byte is reserved for the bracketed-paste
-        // multi-line route (issue #209); we cover that in
-        // [writeInputToPaneWrapsMultiLineInputInBracketedPaste].
-        vm.writeInputToPane("%0", "ls\r".toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(sent.toString(), 2, sent.size)
-        assertEquals("send-keys -l -t %0 -- 'ls'", sent[0])
-        assertEquals("send-keys -t %0 Enter", sent[1])
-    }
-
-    @Test
-    fun writeInputToPaneExitsTmuxCopyModeBeforeSendingBytes() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%0",
-                    windowId = "@0",
-                    sessionId = "\$0",
-                    title = "shell",
-                    paneIndex = 0,
-                    inCopyMode = true,
-                ),
-            ),
-        )
-
-        val result = vm.writeInputToPaneResult("%0", "ls\r".toByteArray(Charsets.UTF_8))
-        runCurrent()
-
-        assertTrue("copy-mode recovery should keep pane input successful", result.isSuccess)
-        assertFalse("copy-mode recovery must not mark tmux disconnected", client.disconnected.value)
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-        assertEquals(
-            listOf(
-                "send-keys -X -t %0 cancel",
-                "send-keys -l -t %0 -- 'ls'",
-                "send-keys -t %0 Enter",
-            ),
-            client.sentCommands.filter { it.startsWith("send-keys") },
-        )
-        assertFalse(vm.panes.value.single { it.paneId == "%0" }.inCopyMode)
-    }
-
-    @Test
-    fun writeInputToPaneSendKeysFailureSurfacesFailedStatus() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient().apply {
-            closeAndThrowOnCommandPrefix = "send-keys"
-            closeAndThrowException = TmuxClientException(
-                "tmux command `send-keys` timed out after 100ms",
-            )
-            closeAndThrowDisconnectEvent = TmuxDisconnectEvent(
-                reason = TmuxDisconnectReason.CommandTimeout,
-                source = "command_timeout",
-                intent = "command_timeout",
-                commandKind = "send-keys",
-                timeoutMode = "fatal",
-            )
-        }
-        vm.attachClientForTest(client)
-        runCurrent()
-
-        vm.writeInputToPane("%0", "ls\r".toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        assertTrue(
-            "expected send-keys dispatch, got ${client.sentCommands}",
-            client.sentCommands.any { it.startsWith("send-keys") },
-        )
-        val status = vm.connectionStatus.value
-        assertTrue(
-            "expected Failed after send-keys failure, got $status",
-            status is TmuxSessionViewModel.ConnectionStatus.Failed,
-        )
-        assertEquals(
-            "Tmux command timed out from test@test:0. Tap Reconnect to retry.",
-            (status as TmuxSessionViewModel.ConnectionStatus.Failed).message,
-        )
-    }
-
-    @Test
     fun eofDisconnectUnregistersDeadClientAndPreservesReconnectTarget() = runTest(scheduler) {
         val registry = ActiveTmuxClients()
         val vm = newVm(registry)
@@ -6079,56 +5982,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
     }
 
     @Test
-    fun writeInputToPaneSeparatesLeadingDashLiteralFromTmuxOptions() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.writeInputToPane("%0", "-tproof".toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        val cmd = client.sentCommands.single { it.startsWith("send-keys") }
-        assertEquals("send-keys -l -t %0 -- '-tproof'", cmd)
-    }
-
-    @Test
-    fun writeInputToPaneEscapesSingleQuotesViaCloseEscapeOpen() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.writeInputToPane("%2", "it's".toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        // Single-quote sequence: closes ', escapes the literal quote, then
-        // reopens. The composer wraps with outer single quotes too.
-        val cmd = client.sentCommands.single { it.startsWith("send-keys") }
-        assertTrue(
-            "expected POSIX-shell-style escape in $cmd",
-            cmd == "send-keys -l -t %2 -- 'it'\\''s'",
-        )
-    }
-
-    @Test
-    fun terminalStateInputRoutesThroughTmuxSendKeys() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "\$0", "shell", paneIndex = 0)),
-        )
-        advanceUntilIdle()
-
-        vm.panes.value.single().terminalState.writeInput("echo ok\r".toByteArray(Charsets.UTF_8))
-        waitForSentCommandCount(client, expectedCount = 2)
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals("send-keys -l -t %0 -- 'echo ok'", sent[0])
-        assertEquals("send-keys -t %0 Enter", sent[1])
-    }
-
-    @Test
     fun codexScaleTmuxOutputFloodKeepsTerminalAndConnectionStateConsistent() =
         runTest(scheduler, timeout = SLOW_FEED_RUN_TEST_TIMEOUT) {
         val diagnostics = installRecordingDiagnosticSink()
@@ -6894,177 +6747,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
         )
     }
 
-    @Test
-    fun dequeuedTmuxInputSendFailureRetriesOnceBeforeRecordingSent() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient().apply {
-            throwOnCommandPrefix = "send-keys -l -t %0"
-            throwOnCommandRemaining = 1
-        }
-        vm.attachClientForTest(client)
-        val sink = vm.tmuxInputSinkForTest("%0")
-
-        sink.write("retry".toByteArray(Charsets.US_ASCII))
-        advanceUntilIdle()
-
-        val literalSends = client.sentCommands.filter { it.startsWith("send-keys -l -t %0") }
-        assertEquals(
-            "the dequeued batch must be retried once after a post-dequeue send failure",
-            2,
-            literalSends.size,
-        )
-        val metrics = vm.tmuxInputMetricsForTest("%0") ?: error("input metrics should be recorded")
-        assertEquals(5L, metrics.totalEnqueuedBytes)
-        assertEquals(
-            "bytes should only be counted sent after the retry succeeds",
-            5L,
-            metrics.totalSentBytes,
-        )
-        assertFalse("one recovered send failure must not disconnect the client", client.disconnected.value)
-    }
-
-    @Test
-    fun tmuxHighRateInputStressBatchesWithBoundedBacklogAndNoContentLoss() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient().apply {
-            sendCommandDelayMs = 3L
-        }
-        vm.attachClientForTest(client)
-        val sink = vm.tmuxInputSinkForTest("%0")
-        val chunks = List(800) { index ->
-            ("stress-${index.toString().padStart(4, '0')}-" + "x".repeat(51))
-                .toByteArray(Charsets.US_ASCII)
-        }
-        val expected = chunks.fold(ByteArray(0)) { acc, bytes -> acc + bytes }
-
-        val writer = Thread({
-            chunks.forEach { sink.write(it) }
-        }, "tmux-input-stress-writer")
-        writer.start()
-
-        waitForTmuxInputBytes(vm, paneId = "%0", expectedBytes = expected.size.toLong(), writer = writer)
-        writer.join(1_000)
-        sink.close()
-
-        val metrics = vm.tmuxInputMetricsForTest("%0")
-            ?: error("stress metrics should be recorded")
-        assertEquals(expected.size.toLong(), metrics.totalEnqueuedBytes)
-        assertEquals(expected.size.toLong(), metrics.totalSentBytes)
-        assertTrue(
-            "stress should build real backlog; metrics=$metrics",
-            metrics.maxPendingBytes > TMUX_INPUT_MAX_BATCH_BYTES,
-        )
-        assertTrue(
-            "backlog must stay within bounded queue capacity; metrics=$metrics",
-            metrics.maxPendingBytes <= vm.tmuxInputCapacityBytesForTest(),
-        )
-        assertTrue("input should be batched; metrics=$metrics", metrics.sentBatchCount < chunks.size)
-        assertTrue("batch size metric should be recorded; metrics=$metrics", metrics.maxBatchBytes > chunks.first().size)
-        assertTrue("batch size must remain bounded; metrics=$metrics", metrics.maxBatchBytes <= TMUX_INPUT_MAX_BATCH_BYTES)
-        assertTrue("send latency metric should be recorded; metrics=$metrics", metrics.maxSendLatencyMs > 0.0)
-        writeTmuxInputStressReport(metrics, expectedBytes = expected.size, chunks = chunks.size)
-
-        val reconstructed = client.sentCommands
-            .filter { it.startsWith("send-keys -l -t %0 -- '") }
-            .joinToString(separator = "") { command ->
-                command.substringAfter("-- '").removeSuffix("'")
-            }
-            .toByteArray(Charsets.US_ASCII)
-        assertEquals(
-            "high-rate stress must not lose or reorder input bytes",
-            expected.toList(),
-            reconstructed.toList(),
-        )
-    }
-
-    @Test
-    fun tmuxInputStreamDropsOverflowAndAcceptsLaterBytes() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient().apply {
-            sendCommandGatePrefix = "send-keys -l -t %0"
-            sendCommandGate = CompletableDeferred()
-        }
-        vm.attachClientForTest(client)
-        val sink = vm.tmuxInputSinkForTest("%0")
-        val chunk = ByteArray(TMUX_INPUT_CHUNK_BYTES) { 'x'.code.toByte() }
-
-        repeat(vm.tmuxInputCapacityBytesForTest() / TMUX_INPUT_CHUNK_BYTES) {
-            sink.write(chunk)
-        }
-        runCurrent()
-
-        val overflow = runCatching { sink.write("overflow".toByteArray(Charsets.US_ASCII)) }
-        assertTrue(
-            "the bridge-facing input stream must drop overflow instead of killing the input drainer",
-            overflow.isSuccess,
-        )
-        client.sendCommandGate?.complete(Unit)
-        advanceUntilIdle()
-
-        sink.write("after".toByteArray(Charsets.US_ASCII))
-        advanceUntilIdle()
-
-        val literalSends = client.sentCommands.filter { it.startsWith("send-keys -l -t %0") }
-        assertTrue(
-            "later input must still reach tmux after a transient full-queue drop",
-            literalSends.any { it.contains("'after'") },
-        )
-        assertTrue(
-            "overflow bytes should be dropped rather than replayed after the backlog drains",
-            literalSends.none { it.contains("overflow") },
-        )
-    }
-
-    @Test
-    fun terminalDaQueryResponsesSuppressedInBridgeMode() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "\$0", "shell", paneIndex = 0)),
-        )
-        advanceUntilIdle()
-
-        val state = vm.panes.value.single().terminalState
-        val queryBytes = "\u001b[c".toByteArray(Charsets.US_ASCII)
-        state.appendRemoteOutput(queryBytes)
-        drainTerminalBridgeHandler()
-
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertTrue(
-            "bridge-mode emulator must not generate DA query responses, got $sent",
-            sent.none { it.contains("send-keys -H") || it.contains("send-keys -l") },
-        )
-    }
-
-    @Test
-    fun terminalOsc11QueryResponsesSuppressedInBridgeMode() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "\$0", "shell", paneIndex = 0)),
-        )
-        advanceUntilIdle()
-
-        val state = vm.panes.value.single().terminalState
-        val queryBytes = "\u001b]11;?\u001b\\".toByteArray(Charsets.US_ASCII)
-        state.appendRemoteOutput(queryBytes)
-        drainTerminalBridgeHandler()
-
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertTrue(
-            "bridge-mode emulator must not generate OSC 11 color query responses, got $sent",
-            sent.none { it.contains("send-keys -H") || it.contains("send-keys -l") },
-        )
-    }
-
     /**
      * Wall-clock-bounded looper pump for the flood-drain tests whose REAL
      * [SshTerminalBridge] feed runs on an Android Handler / background thread
@@ -7106,100 +6788,12 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
             condition = condition,
         )
 
-    private fun TestScope.waitForSentCommandCount(client: FakeTmuxClient, expectedCount: Int) {
-        repeat(100) {
-            advanceUntilIdle()
-            if (client.sentCommands.count { command -> command.startsWith("send-keys") } >= expectedCount) {
-                return
-            }
-            Thread.sleep(10)
-        }
-        advanceUntilIdle()
-        assertTrue(
-            "expected at least $expectedCount send-keys commands, got ${client.sentCommands}",
-            client.sentCommands.count { it.startsWith("send-keys") } >= expectedCount,
-        )
-    }
-
-    private fun TestScope.waitForTmuxInputBytes(
-        vm: TmuxSessionViewModel,
-        paneId: String,
-        expectedBytes: Long,
-        writer: Thread,
-    ) {
-        repeat(10_000) {
-            advanceTimeBy(3L)
-            runCurrent()
-            val metrics = vm.tmuxInputMetricsForTest(paneId)
-            if (metrics?.totalSentBytes == expectedBytes && !writer.isAlive) return
-            Thread.sleep(1)
-        }
-        val metrics = vm.tmuxInputMetricsForTest(paneId)
-        assertEquals(
-            "timed out waiting for tmux input stress drain; metrics=$metrics writerAlive=${writer.isAlive}",
-            expectedBytes,
-            metrics?.totalSentBytes,
-        )
-    }
-
-    private fun writeTmuxInputStressReport(
-        metrics: TmuxInputStressMetrics,
-        expectedBytes: Int,
-        chunks: Int,
-    ) {
-        val outputDir = if (File("settings.gradle.kts").isFile) {
-            File("app/build/reports/tmux-input-stress")
-        } else {
-            File("build/reports/tmux-input-stress")
-        }
-        val report = File(outputDir, "high-rate-input.json")
-        report.parentFile?.mkdirs()
-        report.writeText(
-            """
-            {
-              "stress": "tmux-high-rate-input",
-              "input_chunks": $chunks,
-              "expected_bytes": $expectedBytes,
-              "max_pending_capacity_bytes": $TMUX_INPUT_MAX_PENDING_BYTES,
-              "max_batch_capacity_bytes": $TMUX_INPUT_MAX_BATCH_BYTES,
-              "metrics": {
-                "total_enqueued_bytes": ${metrics.totalEnqueuedBytes},
-                "total_sent_bytes": ${metrics.totalSentBytes},
-                "max_pending_bytes": ${metrics.maxPendingBytes},
-                "max_pending_chunks": ${metrics.maxPendingChunks},
-                "max_batch_bytes": ${metrics.maxBatchBytes},
-                "max_batch_chunks": ${metrics.maxBatchChunks},
-                "sent_batch_count": ${metrics.sentBatchCount},
-                "max_send_latency_ms": ${metrics.maxSendLatencyMs}
-              },
-              "no_content_loss": ${metrics.totalEnqueuedBytes == expectedBytes.toLong() && metrics.totalSentBytes == expectedBytes.toLong()}
-            }
-            """.trimIndent(),
-        )
-    }
-
-    private fun drainTerminalBridgeHandler() {
-        shadowOf(Looper.getMainLooper()).idle()
-    }
-
     private fun renderedTranscriptFrom(state: TerminalSurfaceState): String {
         val bridgeField = TerminalSurfaceState::class.java.getDeclaredField("bridge").apply {
             isAccessible = true
         }
         val bridge = bridgeField.get(state) as? SshTerminalBridge ?: return ""
         return bridge.emulator.screen.transcriptText
-    }
-
-    @Test
-    fun onKeyBarKeyIgnoresUnknownLabel() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.onKeyBarKey("%0", "ZorkKey")
-        advanceUntilIdle()
-
-        assertTrue(client.sentCommands.none { it.startsWith("send-keys") })
     }
 
     // ─── Issue #282: detected agents no longer seed a popup hint ────
@@ -10951,73 +10545,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
             ),
             client.sentCommands.filter { it.startsWith("send-keys") },
         )
-    }
-
-    @Test
-    fun tmuxInputStreamLargeSingleWriteEnqueuesBoundedPrefixInsteadOfDroppingWholePaste() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient().apply {
-            sendCommandGatePrefix = "send-keys -l -t %0"
-            sendCommandGate = CompletableDeferred()
-        }
-        vm.attachClientForTest(client)
-        val sink = vm.tmuxInputSinkForTest("%0")
-        val capacity = vm.tmuxInputCapacityBytesForTest()
-        val payload = ByteArray(capacity + TMUX_INPUT_CHUNK_BYTES) { index ->
-            if (index < capacity) 'a'.code.toByte() else 'z'.code.toByte()
-        }
-
-        val write = runCatching { sink.write(payload) }
-        assertTrue(
-            "a single paste larger than the queue capacity must not be rejected before any bytes are enqueued",
-            write.isSuccess,
-        )
-        val beforeDrain = vm.tmuxInputMetricsForTest("%0")
-            ?: error("input metrics should be recorded")
-        assertEquals(
-            "the bounded queue should retain the prefix it can send instead of dropping the whole paste",
-            capacity.toLong(),
-            beforeDrain.totalEnqueuedBytes,
-        )
-
-        runCurrent()
-        client.sendCommandGate?.complete(Unit)
-        advanceUntilIdle()
-
-        val sentText = client.sentCommands
-            .filter { it.startsWith("send-keys -l -t %0 -- '") }
-            .joinToString(separator = "") { command ->
-                command.substringAfter("-- '").removeSuffix("'")
-            }
-        assertEquals(
-            "only the bounded prefix should be sent after the overflow tail is dropped",
-            capacity,
-            sentText.length,
-        )
-        assertTrue("the retained prefix should reach tmux", sentText.all { it == 'a' })
-        assertFalse("overflow tail bytes must not be replayed later", sentText.any { it == 'z' })
-    }
-
-    @Test
-    fun tmuxInputStreamWriteAfterCloseThrowsInsteadOfSilentlyDroppingBytes() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-        val sink = vm.tmuxInputSinkForTest("%0")
-
-        sink.close()
-        val write = runCatching { sink.write("after-close".toByteArray(Charsets.US_ASCII)) }
-
-        assertTrue(
-            "writes after close must be observable failures, not silent successful drops",
-            write.isFailure,
-        )
-        assertTrue(
-            "closed queue writes should fail with IOException",
-            write.exceptionOrNull() is IOException,
-        )
-        advanceUntilIdle()
-        assertTrue(client.sentCommands.none { it.startsWith("send-keys") })
     }
 
     @Test
